Extract project template selection view
