+++ Nyxian/UI/ContentView.swift
@@ -392,79 +392,6 @@ import UIKit
     }
 }
 
-struct ProjectTemplateSelectionView: View {
-    @ObservedObject var model: ProjectTemplateOptionsModel
-
-    var body: some View {
-        VStack(spacing: 8) {
-            templateRow(
-                title: "App",
-                subtitle: "Application project",
-                systemImage: "app.badge",
-                projectType: .app
-            )
-
-            templateRow(
-                title: "Utility",
-                subtitle: "Command line tool",
-                systemImage: "terminal",
-                projectType: .utility
-            )
-        }
-        .padding(.top, 2)
-        .padding(.horizontal, 18)
-        .padding(.bottom, 6)
-        .fixedSize(horizontal: false, vertical: true)
-    }
-
-    private func templateRow(title: String,
-                             subtitle: String,
-                             systemImage: String,
-                             projectType: NXProjectType) -> some View {
-        let isSelected = model.projectType == projectType
-
-        return Button {
-            model.selectProjectType(projectType)
-        } label: {
-            HStack(spacing: 12) {
-                ZStack {
-                    RoundedRectangle(cornerRadius: 10, style: .continuous)
-                        .fill(isSelected ? Color.accentColor : Color(uiColor: .tertiarySystemFill))
-                    Image(systemName: systemImage)
-                        .font(.system(size: 22, weight: .semibold))
-                        .foregroundStyle(isSelected ? Color.white : Color.accentColor)
-                }
-                .frame(width: 42, height: 42)
-
-                VStack(alignment: .leading, spacing: 2) {
-                    Text(title)
-                        .font(.body.weight(.semibold))
-                        .foregroundStyle(.primary)
-                    Text(subtitle)
-                        .font(.caption)
-                        .foregroundStyle(.secondary)
-                }
-
-                Spacer(minLength: 8)
-
-                Image(systemName: isSelected ? "checkmark.circle.fill" : "circle")
-                    .font(.system(size: 20, weight: .semibold))
-                    .foregroundStyle(isSelected ? Color.accentColor : Color(uiColor: .tertiaryLabel))
-            }
-            .padding(10)
-            .frame(maxWidth: .infinity, alignment: .leading)
-            .background(Color(uiColor: .secondarySystemFill))
-            .clipShape(RoundedRectangle(cornerRadius: 13, style: .continuous))
-            .overlay {
-                RoundedRectangle(cornerRadius: 13, style: .continuous)
-                    .stroke(isSelected ? Color.accentColor : Color.clear, lineWidth: 1.5)
-            }
-        }
-        .buttonStyle(.plain)
-        .accessibilityAddTraits(isSelected ? .isSelected : [])
-    }
-}
-
 final class ProjectTemplateOptionsModel: ObservableObject {
     private static let organizationIdentifierDefaultsKey = "LDEOrganizationPrefix"
     private static let defaultOrganizationIdentifier = "com.example"
